refactor: "studying"
