+++ test/rank.test.js
@@ -1,6 +1,6 @@
 const Rank = require("../service/rank");
 const RankModels = require("../models/rank");
-const StudyingModels = require("../models/studying");
+const StudyingModels = require("../models/studying.model");
 
 describe("user(20대), ageGroup이 전체 일 때", () => {
   const user = {
